Add square, quote cmds, make case-insensitivity job of cmds not message

+++ commands/print.js
@@ -1,7 +1,8 @@
 exports.run = function(message, args) {
-  message.delete();
-  const embed = require('../util/embed.js').run()
-    .setDescription(args.join(' ').slice(7))
-    .setFooter(`${message.author.username}#${message.author.discriminator}`);
-  message.channel.sendEmbed(embed);
+  if (args.length > 1) {
+    message.delete();
+    const embed = require('../util/embed.js').run()
+      .setDescription(args.join(' ').slice(args[0].length));
+    message.channel.sendEmbed(embed);
+  }
 };
